Extract OAuth related files to oauth module
