+++ api-client-oauth/build.gradle
@@ -19,37 +19,18 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-apply plugin: 'com.android.application'
 
-Properties local_properties = new Properties()
-File localPropertiesFile = project.rootProject.file('local.properties')
-if (localPropertiesFile.exists()) {
-    local_properties.load(localPropertiesFile.newDataInputStream())
-}
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion versions.COMPILE_SDK_VERSION
     buildToolsVersion versions.BUILD_TOOLS_VERSION
 
     defaultConfig {
-        applicationId GROUP_ID + ".sample"
-
         minSdkVersion versions.MIN_SDK_VERSION
-        targetSdkVersion versions.TARGET_SDK_VERSION
 
         versionCode = VERSION_CODE
         versionName = VERSION
-
-        /**
-         * === KEYS ===
-         * Define in your local.properties file the next two keys:
-         *      - oauth_consumer_key
-         *      - oauth_consumer_secret
-         *
-         * to obtain them visit https://dev.xing.com/
-         */
-        buildConfigField "String", "OAUTH_CONSUMER_KEY", "\"" + (local_properties.oauth_consumer_key ?: "") + "\""
-        buildConfigField "String", "OAUTH_CONSUMER_SECRET", "\"" + (local_properties.oauth_consumer_secret ?: "") + "\""
     }
 
     buildTypes {
@@ -63,6 +44,11 @@ android {
         disable 'InvalidPackage'
     }
 
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_7
+        targetCompatibility JavaVersion.VERSION_1_7
+    }
+
     packagingOptions {
         exclude 'LICENSE.txt'
         exclude 'META-INF/DEPENDENCIES'
@@ -71,20 +57,10 @@ android {
         exclude 'META-INF/LICENSE.txt'
         exclude 'META-INF/NOTICE.txt'
     }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
-    }
 }
 
 dependencies {
+    provided libraries.supportAnnotations
+    compile libraries.singpost
     compile project(':sdk')
-    compile project(':api-client-oauth')
-    compile libraries.supportV4
-    compile libraries.supportAppCompat
-    compile libraries.supportDesign
-    compile libraries.supportCardView
-    compile libraries.supportRecyclerView
-    compile libraries.glide
 }
