+++ api-client-oauth/src/main/res/layout/activity_oauth_callback.xml
@@ -24,7 +24,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.xing.android.sdk.login.OauthCallbackActivity">
+    tools:context="com.xing.android.api.oauth.OauthCallbackActivity">
     <WebView
         android:id="@+id/webView"
         android:layout_width="match_parent"
